Updated getPeakGain Function
Now the function is able to give the value of Peak Gain for SISO as well as MIMO systems correctly till two places of decimal. Increasing the accuracy further will lead to a greater amount of memory used by the program and increased run time. Therefore the program accuracy is held reasonably to three places of decimal in most cases and till two places of decimal in the worst cases.